requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
Pillow>=10.0
pandas>=2.0
tqdm>=4.60
numpy<2
nbformat>=5.0
